+++ sdk/cosmos/azure_data_cosmos/src/clients/container_client.rs
@@ -3,11 +3,14 @@
 
 use crate::{
     clients::OffersClient,
-    models::{ContainerProperties, CosmosResponse, ThroughputProperties},
+    models::{
+        BatchResponse, ContainerProperties, CosmosResponse, ItemResponse, ResourceResponse,
+        ThroughputProperties,
+    },
     options::{BatchOptions, QueryOptions, ReadContainerOptions},
     pipeline::GatewayPipeline,
     resource_context::{ResourceLink, ResourceType},
-    transactional_batch::{TransactionalBatch, TransactionalBatchResponse},
+    transactional_batch::TransactionalBatch,
     DeleteContainerOptions, FeedItemIterator, ItemOptions, PartitionKey, Query,
     ReplaceContainerOptions, ThroughputOptions,
 };
@@ -106,7 +109,7 @@ impl ContainerClient {
             reason = "The 'options' parameter may be used in the future"
         )]
         options: Option<ReadContainerOptions>,
-    ) -> azure_core::Result<CosmosResponse<ContainerProperties>> {
+    ) -> azure_core::Result<ResourceResponse<ContainerProperties>> {
         let cosmos_request =
             CosmosRequest::builder(OperationType::Read, self.link.clone()).build()?;
         let response: CosmosResponse<ContainerProperties> = self
@@ -114,7 +117,7 @@ impl ContainerClient {
             .send(cosmos_request, Context::default())
             .await?;
 
-        Ok(response)
+        Ok(ResourceResponse::new(response))
     }
 
     /// Updates the indexing policy of the container.
@@ -150,13 +153,14 @@ impl ContainerClient {
             reason = "The 'options' parameter may be used in the future"
         )]
         options: Option<ReplaceContainerOptions>,
-    ) -> azure_core::Result<CosmosResponse<ContainerProperties>> {
+    ) -> azure_core::Result<ResourceResponse<ContainerProperties>> {
         let cosmos_request = CosmosRequest::builder(OperationType::Replace, self.link.clone())
             .json(&properties)
             .build()?;
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(ResourceResponse::new)
     }
 
     /// Reads container throughput properties, if any.
@@ -199,7 +203,7 @@ impl ContainerClient {
         &self,
         throughput: ThroughputProperties,
         options: Option<ThroughputOptions>,
-    ) -> azure_core::Result<CosmosResponse<ThroughputProperties>> {
+    ) -> azure_core::Result<ResourceResponse<ThroughputProperties>> {
         #[allow(
             unused_variables,
             reason = "The 'options' variable may be used in the future"
@@ -214,7 +218,10 @@ impl ContainerClient {
             .expect("service should always return a '_rid' for a container");
 
         let offers_client = OffersClient::new(self.pipeline.clone(), resource_id);
-        offers_client.replace(Context::default(), throughput).await
+        offers_client
+            .replace(Context::default(), throughput)
+            .await
+            .map(ResourceResponse::new)
     }
 
     /// Deletes this container.
@@ -231,12 +238,13 @@ impl ContainerClient {
             reason = "The 'options' parameter may be used in the future"
         )]
         options: Option<DeleteContainerOptions>,
-    ) -> azure_core::Result<CosmosResponse<()>> {
+    ) -> azure_core::Result<ResourceResponse<()>> {
         let cosmos_request =
             CosmosRequest::builder(OperationType::Delete, self.link.clone()).build()?;
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(ResourceResponse::new)
     }
 
     /// Creates a new item in the container.
@@ -274,7 +282,7 @@ impl ContainerClient {
     ///
     /// By default, the newly created item is *not* returned in the HTTP response.
     /// If you want the new item to be returned, set the [`ItemOptions::with_content_response_on_write_enabled()`] option to `true`.
-    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`CosmosResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
+    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`ItemResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
     ///
     /// ```rust,no_run
     /// use azure_data_cosmos::ItemOptions;
@@ -307,7 +315,7 @@ impl ContainerClient {
         partition_key: impl Into<PartitionKey>,
         item: T,
         options: Option<ItemOptions>,
-    ) -> azure_core::Result<CosmosResponse<()>> {
+    ) -> azure_core::Result<ItemResponse<()>> {
         let options = options.clone().unwrap_or_default();
         let excluded_regions = options.excluded_regions.clone();
         let mut cosmos_request =
@@ -321,6 +329,7 @@ impl ContainerClient {
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(ItemResponse::new)
     }
 
     /// Replaces an existing item in the container.
@@ -359,7 +368,7 @@ impl ContainerClient {
     ///
     /// By default, the replaced item is *not* returned in the HTTP response.
     /// If you want the replaced item to be returned, set the [`ItemOptions::with_content_response_on_write_enabled()`] option to `true`.
-    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`CosmosResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
+    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`ItemResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
     ///
     /// ```rust,no_run
     /// use azure_data_cosmos::ItemOptions;
@@ -392,7 +401,7 @@ impl ContainerClient {
         item_id: &str,
         item: T,
         options: Option<ItemOptions>,
-    ) -> azure_core::Result<CosmosResponse<()>> {
+    ) -> azure_core::Result<ItemResponse<()>> {
         let link = self.items_link.item(item_id);
         let options = options.clone().unwrap_or_default();
         let excluded_regions = options.excluded_regions.clone();
@@ -406,6 +415,7 @@ impl ContainerClient {
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(ItemResponse::new)
     }
 
     /// Creates or replaces an item in the container.
@@ -447,7 +457,7 @@ impl ContainerClient {
     ///
     /// By default, the created/replaced item is *not* returned in the HTTP response.
     /// If you want the created/replaced item to be returned, set the [`ItemOptions::with_content_response_on_write_enabled()`] option to `true`.
-    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`CosmosResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
+    /// You can deserialize the returned item by retrieving the [`ResponseBody`](azure_core::http::response::ResponseBody) using [`ItemResponse::into_body`] and then calling [`ResponseBody::json`](azure_core::http::response::ResponseBody::json), like this:
     ///
     /// ```rust,no_run
     /// use azure_data_cosmos::ItemOptions;
@@ -479,7 +489,7 @@ impl ContainerClient {
         partition_key: impl Into<PartitionKey>,
         item: T,
         options: Option<ItemOptions>,
-    ) -> azure_core::Result<CosmosResponse<()>> {
+    ) -> azure_core::Result<ItemResponse<()>> {
         let options = options.clone().unwrap_or_default();
         let excluded_regions = options.excluded_regions.clone();
         let mut cosmos_request =
@@ -493,7 +503,8 @@ impl ContainerClient {
         return self
             .container_connection
             .send(cosmos_request, Context::default())
-            .await;
+            .await
+            .map(ItemResponse::new);
     }
 
     /// Reads a specific item from the container.
@@ -532,7 +543,7 @@ impl ContainerClient {
         partition_key: impl Into<PartitionKey>,
         item_id: &str,
         options: Option<ItemOptions>,
-    ) -> azure_core::Result<CosmosResponse<T>> {
+    ) -> azure_core::Result<ItemResponse<T>> {
         let mut options = options.unwrap_or_default();
 
         // Read APIs should always return the item, ignoring whatever the user set.
@@ -549,6 +560,7 @@ impl ContainerClient {
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(|r| ItemResponse::new(r))
     }
 
     /// Deletes an item from the container.
@@ -577,7 +589,7 @@ impl ContainerClient {
         partition_key: impl Into<PartitionKey>,
         item_id: &str,
         options: Option<ItemOptions>,
-    ) -> azure_core::Result<CosmosResponse<()>> {
+    ) -> azure_core::Result<ItemResponse<()>> {
         let link = self.items_link.item(item_id);
         let options = options.clone().unwrap_or_default();
         let excluded_regions = options.excluded_regions.clone();
@@ -590,6 +602,7 @@ impl ContainerClient {
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(ItemResponse::new)
     }
 
     /// Executes a single-partition query against items in the container.
@@ -726,7 +739,7 @@ impl ContainerClient {
         &self,
         batch: TransactionalBatch,
         options: Option<BatchOptions>,
-    ) -> azure_core::Result<CosmosResponse<TransactionalBatchResponse>> {
+    ) -> azure_core::Result<BatchResponse> {
         let options = options.unwrap_or_default();
         let partition_key = batch.partition_key().clone();
 
@@ -740,5 +753,6 @@ impl ContainerClient {
         self.container_connection
             .send(cosmos_request, Context::default())
             .await
+            .map(BatchResponse::new)
     }
 }