+++ sdk/cosmos/azure_data_cosmos/src/lib.rs
@@ -30,7 +30,7 @@ pub use account_endpoint::CosmosAccountEndpoint;
 pub use account_reference::CosmosAccountReference;
 pub use connection_string::*;
 pub use credential::CosmosCredential;
-pub use models::CosmosResponse;
+pub use models::{BatchResponse, CosmosDiagnostics, ItemResponse, ResourceResponse};
 pub use options::*;
 pub use partition_key::*;
 pub use query::Query;
@@ -40,7 +40,7 @@ pub use transactional_batch::{
     TransactionalBatch, TransactionalBatchOperationResult, TransactionalBatchResponse,
 };
 
-pub use feed::{FeedItemIterator, FeedPage, FeedPageIterator};
+pub use feed::{FeedItemIterator, FeedPage, FeedPageIterator, QueryFeedPage};
 mod background_task_manager;
 mod cosmos_request;
 #[cfg(feature = "fault_injection")]